+++ apps/docs/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @orderly/docs
 
+## 1.1.1
+
+### Patch Changes
+
+- publish
+- 0b52044b6: publish
+- Updated dependencies
+- Updated dependencies [0b52044b6]
+  - @orderly.network/react@2.1.1
+  - @orderly.network/hooks@2.1.1
+  - @orderly.network/web3-onboard@2.1.1
+  - @orderly.network/types@2.1.1
+
+## 1.1.1-alpha.0
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/react@2.1.1-alpha.0
+  - @orderly.network/hooks@2.1.1-alpha.0
+  - @orderly.network/web3-onboard@2.1.1-alpha.0
+  - @orderly.network/types@2.1.1-alpha.0
+
 ## 1.1.0
 
 ### Minor Changes